BS_2630 SOLVE
