+++ BOSEONG/WEEK5/BOSEONG/BS_1992.java
@@ -1,5 +1,7 @@
 package BOSEONG.WEEK5.BOSEONG;
 
 public class BS_1992 {
-    
+    public static void main(String[] args) {
+        
+    }
 }
